pyproject.toml
[tool.poetry]
name = "desafio_tikal"
version = "0.1.0"
description = ""
authors = ["CaMMelo <cmelo.marcelo.pereira@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.12.4"
Markdown = "^3.3.4"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
isort = "^5.8.0"
black = "^20.8b1"
flake8 = "^3.9.1"
factory-boy = "^3.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"